+++ CHANGELOG.md
@@ -37,8 +37,20 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+### State Machine Breaking
+
+(claims) [\#605](https://github.com/tharsis/evmos/pull/605) Remove duplicated `SetClaimsRecord`.
+
+### API Breaking
+
+(claims) [\#605](https://github.com/tharsis/evmos/pull/605) Remove `claims-` prefix in CLI query commands.
+(erc20) [\#592](https://github.com/tharsis/evmos/pull/592) Finish module completeness audit.
+
+## [v4.0.1] - 2022-05-10
+
+### Bug Fixes
+
 (erc20) [\#588](https://github.com/tharsis/evmos/pull/588) Revert PR [\#556](https://github.com/tharsis/evmos/pull/556).
-(erc20) [\#592](https://github.com/tharsis/evmos/pull/592) Completeness audit
 
 ## [v4.0.0] - 2022-05-09